Refactored for preprocessing grammars
Not yet implemented, but data files are buffered and only grammar files
are processed in a first stage. Then data files are processed in the
flushing stage. But for this to work as intended (that is: with mixed
inputs), parser files have to be required afresh: throws an error for
now.

Also removed shoe-horning streams into buffers to conform to gulp plugin
guidelines.